+++ CHANGELOG.rst
@@ -8,9 +8,10 @@ Current
     - Modified function to read radar traces. 
     - Bug fix for Function `frequency`
     - New functions:
-        - `msg_subset`: For deriving a subset of original messages satisfying given criteria
+        - `msg_subset`: For deriving a subset of original messages satisfying given criteria. See commit #819f2d6
         - `time_subset`: For getting time-slices of original messages satisfying given criteria
         - `acc_state`: Get the ACC (Adaptive Cruise Control) state of driving
+        - `accel_x`: Get the longitudinal acceleration. See commit #819f2d6
 
 - Function `ts_sync` modified to inherit sample from one of the two input dataframes.
 - New function `time_index` and `timeslices` for improved analysis of timeseries data