+++ CHANGELOG.rst
@@ -2,8 +2,8 @@
 Changelog
 =========
 
-Current
----------
+0.1.6 - 2020-Apr-22
+-----------------------
 - class :code:`strymread`
     - Modified function to read radar traces. 
     - Bug fix for Function `frequency`
@@ -12,6 +12,7 @@ Current
         - `time_subset`: For getting time-slices of original messages satisfying given criteria
         - `acc_state`: Get the ACC (Adaptive Cruise Control) state of driving
         - `accel_x`: Get the longitudinal acceleration. See commit #819f2d6
+- Address issue #4, and issue #5.
 
 - Function `ts_sync` modified to inherit sample from one of the two input dataframes.
 - New function `time_index` and `timeslices` for improved analysis of timeseries data